1-print_numbers.c

+++ variadic_functions/1-print_numbers.c
@@ -15,7 +15,7 @@ void print_numbers(const char *separator, const unsigned int n, ...)
 	for (i = 0; i < (int)n; i++)
 	{
 		printf("%d", va_arg(ap, int));
-		if (i < n - 1 && separator)
+		if (i < (int)n - 1 && separator)
 			printf("%s", separator);
 	}
 	printf("\n");
